refactor(sdk): consolidate reflection helper types
